Refactor inspection tab view for designer
