+++ gui/BrakeDiscInspector_GUI_ROI/Workflow/InspectionDatasetTabsControl.xaml.cs
@@ -1,11 +1,6 @@
 using System;
-using System.Diagnostics;
-using System.IO;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Input;
-using BrakeDiscInspector_GUI_ROI.Models;
-using BrakeDiscInspector_GUI_ROI.Util;
 
 namespace BrakeDiscInspector_GUI_ROI.Workflow
 {
@@ -19,66 +14,14 @@ namespace BrakeDiscInspector_GUI_ROI.Workflow
             InitializeComponent();
         }
 
-        private void LoadModelButton_Click(object sender, RoutedEventArgs e)
+        private void InspectionTab_LoadModelRequested(object sender, LoadModelRequestedEventArgs e)
         {
-            var roiIndex = 0;
-
-            if (sender is FrameworkElement element)
-            {
-                if (element.Tag is int intTag)
-                {
-                    roiIndex = intTag;
-                }
-                else if (element.Tag is string strTag && int.TryParse(strTag, out var parsed))
-                {
-                    roiIndex = parsed;
-                }
-                else if (element.DataContext is InspectionRoiConfig roi)
-                {
-                    roiIndex = roi.Index;
-                }
-            }
-
-            if (roiIndex <= 0 && sender is FrameworkElement { DataContext: InspectionRoiConfig ctx })
-            {
-                roiIndex = ctx.Index;
-            }
-
-            if (roiIndex <= 0)
-            {
-                return;
-            }
-
-            LoadModelRequested?.Invoke(this, new LoadModelRequestedEventArgs(roiIndex));
+            LoadModelRequested?.Invoke(this, e);
         }
 
-        private void BtnToggleEdit_Click(object sender, RoutedEventArgs e)
+        private void InspectionTab_ToggleEditRequested(object sender, ToggleEditRequestedEventArgs e)
         {
-            if (sender is Button button && button.DataContext is InspectionRoiConfig cfg)
-            {
-                GuiLog.Info($"[workflow-edit] ToggleEditRequested roi='{cfg.Id}' index={cfg.Index} enabled={cfg.Enabled} isEditable={cfg.IsEditable}");
-
-                ToggleEditRequested?.Invoke(this, new ToggleEditRequestedEventArgs(cfg.Id, cfg.Index));
-            }
-            else
-            {
-                GuiLog.Warn("[workflow-edit] BtnToggleEdit_Click ignored: invalid sender/DataContext");
-            }
-        }
-
-        private void DatasetImage_Click(object sender, MouseButtonEventArgs e)
-        {
-            if (sender is Image image && image.DataContext is DatasetPreviewItem item && File.Exists(item.Path))
-            {
-                try
-                {
-                    Process.Start(new ProcessStartInfo(item.Path) { UseShellExecute = true });
-                }
-                catch (Exception ex)
-                {
-                    GuiLog.Warn($"[dataset] Failed to open '{item.Path}': {ex.Message}");
-                }
-            }
+            ToggleEditRequested?.Invoke(this, e);
         }
     }
 }
